// src/interfaces/lobby.d.ts
export interface IMember {
    id: Long;
    hero_id: any;
    team: number;
    name: string;
    slot: number;
    party_id: any;
    meta_level: any;
    meta_xp: any;
    meta_xp_awarded: any;
    leaver_status: number;
    leaver_actions: any;
    channel: number;
    disabled_hero_id: Long[];
    partner_account_type: number;
    enabled_hero_id: Long[];
    coach_team: number;
    coached_account_ids: Long[];
    coach_rating: number;
    pwrd_cyber_cafe_id: Long;
    pwrd_cyber_cafe_name: string;
    disabled_hero_random_bits: any[];
    rank_change: any;
    cameraman: any;
    custom_game_product_ids: any[];
    search_match_type: number;
    favourite_team_packed: Long;
    is_plus_subscriber: boolean;
    rank_tier_updated: any;
    lane_selection_flags: any;
    can_earn_rewards: any;
    live_spectator_team: number;
    was_mvp_last_game: boolean;
    pending_awards: any[];
    pending_awards_on_victory: any[];
    rank_mmr_boost_type: number;
    queue_point_adjustment: any;
    rank_tier: any;
    title: any;
    guild_id: any;
    reports_available: any;
}

export interface ILobbyUpdate {
    lobby_id: Long;
    v2_members: IMember[];
    v2_left_members: IMember[];
    all_members: IMember[];
    member_indices: number[];
    left_member_indices: number[];
    free_member_indices: number[];
    leader_id: Long;
    server_id: Long;
    game_mode: number;
    pending_invites: any[];
    state: number;
    connect: any;
    lobby_type: number;
    allow_cheats: boolean;
    fill_with_bots: boolean;
    intro_mode: boolean;
    game_name: string;
    team_details: any[];
    tutorial_lesson: any;
    tournament_id: Long;
    tournament_game_id: Long;
    server_region: number;
    game_state: number;
    num_spectators: any;
    cm_pick: number;
    match_id: Long;
    allow_spectating: boolean;
    bot_difficulty_radiant: number;
    game_version: number;
    timed_reward_details: any[];
    pass_key: string;
    leagueid: number;
    penalty_level_radiant: number;
    penalty_level_dire: number;
    load_game_id: Long;
    series_type: number;
    radiant_series_wins: number;
    dire_series_wins: number;
    loot_generated: any;
    loot_awarded: any;
    allchat: boolean;
    dota_tv_delay: number;
    custom_game_mode: any;
    custom_map_name: any;
    custom_difficulty: any;
    lan: boolean;
    broadcast_channel_info: any[];
    first_leaver_accountid: Long;
    series_id: Long;
    low_priority: any;
    extra_messages: any[];
    save_game: any;
    first_blood_happened: any;
    match_outcome: number;
    mass_disconnect: any;
    custom_game_id: Long;
    custom_min_players: any;
    custom_max_players: any;
    partner_type: number;
    visibility: number;
    custom_game_crc: any;
    custom_game_auto_created_lobby: any;
    custom_game_timestamp_any;
    previous_series_matches: any;
    previous_match_override: Long;
    custom_game_uses_account_records: any;
    game_start_time: any;
    pause_setting: number;
    lobby_mvp_account_id: Long;
    weekend_tourney_division_id: Long;
    weekend_tourney_skill_level: any;
    weekend_tourney_bracket_round: any;
    bot_difficulty_dire: number;
    bot_radiant: Long;
    bot_dire: Long;
    event_progression_enabled: any[];
    selection_priority_rules: number;
    series_previous_selection_priority_team_id: Long;
    series_current_selection_priority_team_id: Long;
    series_current_priority_team_choice: number;
    series_current_non_priority_team_choice: number;
    series_current_selection_priority_used_coin_toss: any;
    current_primary_event: number;
    emergency_disabled_hero_ids: Long[];
    custom_game_private_key: any;
    custom_game_penalties: any;
    lan_host_ping_location: any;
    league_node_id: number;
    match_duration: any;
    custom_game_browseable: any;
    league_phase: number;
    record_detailed_stats: any;
    experimental_gameplay_enabled: any;
    guild_challenges: any[];
    guild_details: any[];
    lobby_event_points: any[];
    requested_hero_ids: any[];
    coach_friend_requests: any[];
}

export interface IPlayer {
    id: number;
    team: number;
    nickname: string;
}